reorganised the entire backend server directory as well as refacotring some code
