+++ .gitignore
@@ -14,6 +14,7 @@ pids
 
 # Dependency directories
 node_modules/
+package-lock.json
 
 # Coverage directory used by tools like istanbul
 coverage/
