+++ server/config/awsConfig.js
@@ -1,0 +1,9 @@
+const aws = require("aws-sdk");
+
+aws.config.update({
+  accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+  secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+  region: process.env.AWS_REGION,
+});
+
+module.exports = aws;
