+++ server/config/fieldMapping.js
@@ -1,0 +1,6 @@
+const fieldMapping = {
+  "Given Name(s)": "form1[0].#subform[0].Line1_GivenName[0]",
+  Surname: "form1[0].#subform[0].Line1_FamilyName[0]",
+};
+
+module.exports = fieldMapping;
