--- server/models/conversationModel.js
@@ -1,16 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const conversationSchema = new Schema({
-  user: {
-    type: mongoose.Schema.Types.ObjectId,
-    ref: 'User',
-    required: true,
-  },
-  startTime: {
-    type: Date,
-    default: Date.now(),
-  }
-});
-
-module.exports = mongoose.model("Conversation", conversationSchema);
