--- server/models/messageModel.js
@@ -1,24 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const messageSchema = new Schema({
-  conversation: {
-    type: mongoose.Schema.Types.ObjectId,
-    ref: "Conversation",
-    required: true,
-  },
-  sender: {
-    type: String,
-    required: true,
-  },
-  content: {
-    type: String,
-    required: true,
-  },
-  timestamp: {
-    type: Date,
-    default: Date.now,
-  },
-});
-
-module.exports = mongoose.model("Message", messageSchema);
