+++ server/services/emailService.js
@@ -1,0 +1,14 @@
+const transporter = require('../config/nodemailerConfig');
+
+const sendEmail = (mailOptions) => {
+  return new Promise((resolve, reject) => {
+    transporter.sendMail(mailOptions, (error, info) => {
+      if (error) {
+        reject(error);
+      }
+      resolve(info);
+    });
+  });
+};
+
+module.exports = { sendEmail };
